Refactor states.
