+++ src/states/base_state.py
@@ -1,0 +1,6 @@
+class RobotState:
+    def __init__(self, robot):
+        self.robot = robot
+
+    def handle(self):
+        raise NotImplementedError
