+++ src/states/enter_taking_line_follower.py
@@ -1,0 +1,25 @@
+from src.states.exceptions import ChangingStateExcaption
+from src.states.line_follower import StateLineFollower
+
+
+class StateEnterTakingLineFollower(StateLineFollower):
+    def __init__(self, robot):
+        super().__init__(robot)
+        self.def_speed = 250
+        self.color_l = 0
+        self.color_r = 0
+        self.actual_regulation = 0
+        T_crit = 1.7
+        K_crit = 5.5
+        self.pid_ = pid.PID(0.6 * K_crit, 0.2 * T_crit, 0.7 * T_crit)
+
+    def check_changing_state(self):
+        colors = ('unknown', 'black', 'blue', 'green', 'yellow', 'red', 'white', 'brown')
+        self.robot.color_sensor_l.mode = 'COL-COLOR'
+        self.robot.color_sensor_r.mode = 'COL-COLOR'
+        if colors[self.robot.color_sensor_l.value()] == 'red' and colors[self.robot.color_sensor_l.value()] == 'red':
+            self.robot.color_sensor_l.mode = 'COL-REFLECT'
+            self.robot.color_sensor_r.mode = 'COL-REFLECT'
+            raise ChangingStateExcaption("INSIDE_TAKING_FIELD")
+        self.robot.color_sensor_l.mode = 'COL-REFLECT'
+        self.robot.color_sensor_r.mode = 'COL-REFLECT'
