+++ src/states/exceptions.py
@@ -1,0 +1,2 @@
+class ChangingStateExcaption(Exception):
+    pass
