+++ src/states/find_leaving_line.py
@@ -1,0 +1,16 @@
+from src.states.base_state import RobotState
+from src.states.exceptions import ChangingStateExcaption
+
+
+class StateFindLeavingLine(RobotState):
+    def __init__(self, robot):
+        super().__init__(robot)
+
+    def handle(self):
+        """
+            obrot w lewo o 90 stopni i podjechać trochę do przodu
+        """
+        self.robot.drive_forward(500, 250)
+        self.robot.rotate_right_angle_left()
+        self.robot.drive_forward(750, 250)
+        raise ChangingStateExcaption("ENTER_LEAVING_LINE_FOLLOWER")
